+++ src/App.js
@@ -3,6 +3,7 @@ import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import Landing from './view/Landing/Landing';
 import Shopping from './view/Shopping/Shopping';
 import Header from "./component/Header";
+import Order from './component/Order/Order';
 import './assets/style.css';
 import './App.css'
 
@@ -14,6 +15,7 @@ function App() {
         <Routes>
           <Route path="/landing" element={<Landing />} />
           <Route path="/compra/:categoryName" element={<Shopping />} />
+          <Route path="/order" element={<Order />} />
         </Routes>
       </div>
     </Router>
